rename simulation-handler file to simulations
